Adds generator test
