+++ packages/unmock-core/src/__tests__/backend/state.test.ts
@@ -1,11 +1,13 @@
 import axios from "axios";
-import { isEqual } from "lodash";
 import * as path from "path";
 import { Service, UnmockPackage } from "../..";
 import NodeBackend from "../../backend";
-import { includeCodes, changeToConst, responseBody, Arr, changeSingleSchema, changeMinItems, changeMaxItems, changeRequiredStatus } from "openapi-refinements";
+import { Arr } from "openapi-refinements";
 import { OpenAPIObject } from "loas3/dist/generated/full";
+import { gen } from "../../generator-utils";
 import { ISerializedRequest } from "../../interfaces";
+
+const { withCodes, responseBody, noopThrows, compose, times } = gen;
 
 const servicesDirectory = path.join(__dirname, "..", "__unmock__");
 
@@ -15,17 +17,20 @@ describe("Node.js interceptor", () => {
     const unmock = new UnmockPackage(nodeInterceptor);
     let petstore: Service;
     let filestackApi: Service;
+    let slack: Service;
 
     beforeAll(() => {
       unmock.on();
       petstore = unmock.services.petstore;
       filestackApi = unmock.services.filestackApi;
+      slack = unmock.services.slack;
     });
     afterAll(() => unmock.off());
 
     beforeEach(() => {
       petstore.reset();
       filestackApi.reset();
+      slack.reset();
     });
 
     test("t throws when asking for non existing method/path", async () => {
@@ -39,7 +44,7 @@ describe("Node.js interceptor", () => {
     });
 
     test("t gets correct code upon request without other state", async () => {
-      petstore.state((_, o) => includeCodes(true, true, ["200"])(o));
+      petstore.state(withCodes(200));
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.status).toBe(200);
       expect(
@@ -51,10 +56,10 @@ describe("Node.js interceptor", () => {
     });
 
     test("t gets correct state after setting state with status code", async () => {
-      petstore.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeToConst(5)(responseBody(true), [Arr, "id"]),
-      ].reduce((a, b) => b(a), o));
+      petstore.state(
+        withCodes(200),
+        responseBody({ address: [Arr, "id"] }).const(5),
+      );
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.status).toBe(200);
       expect(response.data.every((pet: any) => pet.id === 5)).toBeTruthy();
@@ -64,25 +69,24 @@ describe("Node.js interceptor", () => {
     // there is no message field on 200...
     // also, we need to be explicit about 200 as there is also a default response
     test("t gets correct state after setting state without status code", async () => {
-      petstore.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeSingleSchema(__ => ___ => ({
+      petstore.state(
+        withCodes(200),
+        responseBody({ path: "/pets" }).schema({
           type: "object",
           required: ["message"],
-          properties: { message: { type: "string", enum: ["Hello World"]}}}))
-          (responseBody("/pets", true, ["200"]), []),
-      ].reduce((a, b) => b(a), o));
+          properties: { message: { type: "string", enum: ["Hello World"]}}}),
+      );
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.status).toBe(200);
       expect(response.data.message).toEqual("Hello World");
     });
 
     test("t gets correct state after multiple overriden state requests", async () => {
-      petstore.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeToConst(5)(responseBody(new RegExp("/pets")), [Arr, "id"]),
-        changeToConst(-1)(responseBody(new RegExp("/pets/{[a-zA-Z0-9/]+}")), ["id"]),
-      ].reduce((a, b) => b(a), o));
+      petstore.state(
+        withCodes(200),
+        responseBody({ path: "/pets", address: [Arr, "id"] }).const(5),
+        responseBody({ path: /\/pets\/{[a-zA-Z0-9/]+}/, address: ["id"] }).const(-1),
+      );
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.status).toBe(200);
       expect(response.data.every((pet: any) => pet.id === 5)).toBeTruthy();
@@ -92,20 +96,20 @@ describe("Node.js interceptor", () => {
     });
 
     test("t gets correct state when setting textual response", async () => {
-      filestackApi.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeToConst("foo")(responseBody(new RegExp("[a-zA-Z0-9/]*"), true, ["200"], ["text/plain"]), []),
-      ].reduce((a, b) => b(a), o));
+      filestackApi.state(
+        withCodes(200),
+        responseBody().const("foo"),
+      );
       const response = await axios("https://cloud.filestackapi.com/prefetch");
       expect(response.status).toBe(200);
       expect(response.data).toBe("foo");
     });
 
     test("gets correct state with query parameter when setting textual response", async () => {
-      filestackApi.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeToConst("foo")(responseBody(new RegExp("[a-zA-Z0-9/]*"), true, ["200"], ["text/plain"]), []),
-      ].reduce((a, b) => b(a), o));
+      filestackApi.state(
+        withCodes(200),
+        responseBody().const("foo"),
+      );
       const response = await axios(
         "https://cloud.filestackapi.com/prefetch?apikey=fake",
       );
@@ -114,20 +118,20 @@ describe("Node.js interceptor", () => {
     });
 
     test("t gets correct state when setting textual response with path", async () => {
-      filestackApi.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeToConst("bar")(responseBody(new RegExp("[a-zA-Z0-9/]*"), true, ["200"], ["text/plain"]), []),
-      ].reduce((a, b) => b(a), o));
+      filestackApi.state(
+        withCodes(200),
+        responseBody({ path: "/prefetch" }).const("bar"),
+      );
       const response = await axios("https://cloud.filestackapi.com/prefetch");
       expect(response.status).toBe(200);
       expect(response.data).toBe("bar");
     });
 
     test("t default response turns into 500", async () => {
-      filestackApi.state((_, o) => [
-        includeCodes(true, true, ["default"]),
-        changeToConst("foo")(responseBody(new RegExp("[a-zA-Z0-9/]*"), true, ["default"], ["text/plain"]), []),
-      ].reduce((a, b) => b(a), o));
+      filestackApi.state(
+        withCodes("default"),
+        responseBody().const("foo"),
+      );
       try {
         await axios("https://cloud.filestackapi.com/prefetch");
         throw new Error("Expected a 500 response");
@@ -138,20 +142,21 @@ describe("Node.js interceptor", () => {
     });
 
     test("t sets an entire response from function", async () => {
-      petstore.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        changeToConst([{id: 1, name: "Fluffy"}])(responseBody("/pets", true, ["200"]), []),
-      ].reduce((a, b) => b(a), o));
+      petstore.state(
+        withCodes(200),
+        responseBody({ path: "/pets" }).const([{id: 1, name: "Fluffy"}]),
+      );
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.data).toEqual([{id: 1, name: "Fluffy"}]);
     });
 
     test("t sets an entire response from with request object", async () => {
-      petstore.state((req, o) => [
-        includeCodes(true, true, ["200"]),
-        changeSingleSchema(__ => ___ => ({ type: "string", enum: [req.host]}))
-          (responseBody("/pets", true, ["200"]), []),
-      ].reduce((a, b) => b(a), o));
+      petstore.state(
+        withCodes(200),
+        (req: ISerializedRequest, o: OpenAPIObject) =>
+          responseBody({ path: "/pets" })
+            .schema({ type: "string", enum: [req.host]})(req, o),
+      );
       const response = await axios("http://petstore.swagger.io/v1/pets");
       expect(response.data).toBe("petstore.swagger.io");
     });
@@ -192,18 +197,11 @@ describe("Node.js interceptor", () => {
 
     // this is just a no-op in the new version
     test("t fails setting an array size for non-array elements", async () => {
-      const throwIfUnchanged = (f: (o: OpenAPIObject) => OpenAPIObject) => (o: OpenAPIObject): OpenAPIObject => {
-        const out = f(o);
-        if (isEqual(out, o)) {
-          throw Error("Array item setting did not work");
-        }
-        return out;
-      };
-      petstore.state((_, o) => [
-        includeCodes(true, true, ["200"]),
-        throwIfUnchanged(changeMinItems(5)(responseBody(true), [Arr, "id"])),
-        throwIfUnchanged(changeMaxItems(5)(responseBody(true), [Arr, "id"])),
-      ].reduce((a, b) => b(a), o));
+      petstore.state(
+        withCodes(200),
+        noopThrows(responseBody({ path: "/pets", address: [Arr, "id"] }).minItems(5)),
+        noopThrows(responseBody({ path: "/pets", address: [Arr, "id"] }).maxItems(5)),
+      );
       try {
         await axios("http://petstore.swagger.io/v1/pets");
       } catch (err) {
@@ -217,17 +215,12 @@ describe("Node.js interceptor", () => {
         axios.post("https://slack.com/api/chat.postMessage", {
           data: { channel: "my_channel_id", text },
         });
-      const { slack } = unmock.services;
-      const counter = { n: 0 };
-      slack.state((_, o) => {
-        counter.n += 1;
-        return  [
-          includeCodes(true, true, ["200"]),
-          counter.n <= 3
-          ? changeToConst(text)(responseBody("/chat.postMessage"), ["message", "text"])
-          : (i: OpenAPIObject) => i,
-        ].reduce((a, b) => b(a), o);
-      });
+      slack.state(
+        withCodes(200),
+        times(3)(responseBody({
+          path: "/chat.postMessage",
+          address: ["message", "text"],
+        }).const(text)));
       let resp = await postMessage();
 
       expect(resp.data.message.text).toEqual(text);
@@ -240,12 +233,12 @@ describe("Node.js interceptor", () => {
     });
 
     test("handles 'properties' keyword correctly", async () => {
-      const makeTransformer = (isCat: boolean) => (_: ISerializedRequest, o: OpenAPIObject) =>  [
-        includeCodes(true, true, ["200"]),
-        changeRequiredStatus("properties")(responseBody(new RegExp("/pets/{[a-zA-Z0-9/]+}")), []),
-        changeRequiredStatus("isCat")(responseBody(new RegExp("/pets/{[a-zA-Z0-9/]+}")), ["properties"]),
-        changeToConst(isCat)(responseBody(new RegExp("/pets/{[a-zA-Z0-9/]+}")), ["properties", "isCat"]),
-      ].reduce((a, b) => b(a), o);
+      const makeTransformer = (isCat: boolean) => compose(
+        withCodes(200),
+        responseBody().required("properties"),
+        responseBody({address: ["properties"]}).required("isCat"),
+        responseBody({address: ["properties", "isCat"]}).const(isCat),
+      );
       petstore.state(makeTransformer(true));
       let resp = await axios("http://petstore.swagger.io/v1/pets/54");
       expect(resp.data.properties.isCat).toBeTruthy();
